[add/style] 상세페이지 컴포넌트 분리 및 스타일 작업
[add] 상세페이지 헤더랑 메인 컴포넌트 분리 작업
[style] 상세페이지 댓글 목록 css 작업중
