+++ src/app/(providers)/(default)/projects/[id]/_components/FooterList.tsx
@@ -1,0 +1,13 @@
+import React from "react"
+
+const FooterList = () => {
+  return (
+    <section className="border-2 border-blue-500">
+      <p>이전게시물</p>
+      <p>다음게시물</p>
+      <button>목록</button>
+    </section>
+  )
+}
+
+export default FooterList
